// src/types/SimpleTableConfig.ts
import HeaderObject, { Accessor } from "./HeaderObject";
import Row from "./Row";
import {
  VanillaEmptyStateRenderer,
  VanillaErrorStateRenderer,
  VanillaLoadingStateRenderer,
} from "./RowStateRendererProps";
import FooterRendererProps from "./FooterRendererProps";
import { VanillaHeaderDropdown } from "./HeaderDropdownProps";
import SortColumn, { SortDirection } from "./SortColumn";
import CellClickProps from "./CellClickProps";
import CellChangeProps from "./CellChangeProps";
import { ColumnVisibilityState } from "./ColumnVisibilityTypes";
import { TableFilterState } from "./FilterTypes";
import OnNextPage from "./OnNextPage";
import OnRowGroupExpandProps from "./OnRowGroupExpandProps";
import RowSelectionChangeProps from "./RowSelectionChangeProps";
import { RowButton } from "./RowButton";
import Theme from "./Theme";
import { CustomThemeProps } from "./CustomTheme";
import { GetRowId } from "./GetRowId";
import { ColumnEditorConfig } from "./ColumnEditorConfig";
import { VanillaIconsConfig } from "./IconsConfig";
import { QuickFilterConfig } from "./QuickFilterTypes";

export interface SimpleTableConfig {
  autoExpandColumns?: boolean;
  canExpandRowGroup?: (row: Row) => boolean;
  cellUpdateFlash?: boolean;
  className?: string;
  columnBorders?: boolean;
  columnEditorConfig?: ColumnEditorConfig;
  columnEditorText?: string;
  columnReordering?: boolean;
  columnResizing?: boolean;
  copyHeadersToClipboard?: boolean;
  customTheme?: CustomThemeProps;
  defaultHeaders: HeaderObject[];
  editColumns?: boolean;
  editColumnsInitOpen?: boolean;
  emptyStateRenderer?: VanillaEmptyStateRenderer;
  enableHeaderEditing?: boolean;
  enableRowSelection?: boolean;
  enableStickyParents?: boolean;
  errorStateRenderer?: VanillaErrorStateRenderer;
  expandAll?: boolean;
  externalFilterHandling?: boolean;
  externalSortHandling?: boolean;
  footerRenderer?: (props: FooterRendererProps) => HTMLElement | string | null;
  headerDropdown?: VanillaHeaderDropdown;
  height?: string | number;
  hideFooter?: boolean;
  hideHeader?: boolean;
  icons?: VanillaIconsConfig;
  includeHeadersInCSVExport?: boolean;
  initialSortColumn?: string;
  initialSortDirection?: SortDirection;
  isLoading?: boolean;
  loadingStateRenderer?: VanillaLoadingStateRenderer;
  maxHeight?: string | number;
  onCellClick?: (props: CellClickProps) => void;
  onCellEdit?: (props: CellChangeProps) => void;
  onColumnOrderChange?: (newHeaders: HeaderObject[]) => void;
  onColumnSelect?: (header: HeaderObject) => void;
  onColumnVisibilityChange?: (visibilityState: ColumnVisibilityState) => void;
  onColumnWidthChange?: (headers: HeaderObject[]) => void;
  onFilterChange?: (filters: TableFilterState) => void;
  onGridReady?: () => void;
  onHeaderEdit?: (header: HeaderObject, newLabel: string) => void;
  onLoadMore?: () => void;
  onNextPage?: OnNextPage;
  onPageChange?: (page: number) => void | Promise<void>;
  onRowGroupExpand?: (props: OnRowGroupExpandProps) => void | Promise<void>;
  onRowSelectionChange?: (props: RowSelectionChangeProps) => void;
  onSortChange?: (sort: SortColumn | null) => void;
  quickFilter?: QuickFilterConfig;
  rowButtons?: RowButton[];
  rowGrouping?: Accessor[];
  getRowId?: GetRowId;
  rows: Row[];
  rowsPerPage?: number;
  selectableCells?: boolean;
  selectableColumns?: boolean;
  serverSidePagination?: boolean;
  shouldPaginate?: boolean;
  tableEmptyStateRenderer?: HTMLElement | string | null;
  theme?: Theme;
  totalRowCount?: number;
  useHoverRowBackground?: boolean;
  useOddColumnBackground?: boolean;
  useOddEvenRowBackground?: boolean;
}